scrap_ANAC/requirements.txt
# Dependencias específicas para scrap_ANAC
requests==2.32.4
urllib3==2.4.0
pandas==2.3.0
PyMySQL==1.1.1
SQLAlchemy==2.0.41
google-api-python-client==2.172.0
google-auth==2.40.3
python-dotenv==1.1.0
openpyxl==3.1.5
python-dateutil==2.9.0.post0